Tested with hardware prototyping
Tested all modules except serial comm in conjunction with hardware
prototype. There was a problem in pushBox() where it was using delay()
but was being called by interrupt function. delay() does not work inside
ISR's and so had to switch to sloppy implementation of a bunch of
continuous delayMicroseconds().
The relay NC1 may be blown out due to crossing wires. Will either keep
NO or switch to NC2.